Add non-interactivity to Nextjs install
This update aids in using curl to install the setup script easier. Since there was not a prompt for users, once it got to the install nextJs section, the curl command would complete before a user could complete it. The fix was to add the non-interactive install to create nextjs